Fix template line breaks
Some AI agents (Claude Code for one) will strip out the two
spaces at the end of line used for force a line break in the markdown
templates., meaning line breaks are lost.

Use \ at the end of the line instead to force a line break. \ for line breaks
is supported by most markdown previewers - it's not quite as standard
as the two spaces but it works everywhere that most people care about
(Git Hub, VS Code markdown previewer, etc.)